requirements.txt
numpy
pandas
python-dotenv
psycopg2
scikit-learn
seaborn
catboost
xgboost
# -e .